refactor: isolate client update plan construction
