+++ loader/core/src/main/java/pl/skidam/automodpack_loader_core/client/ModpackUpdater.java
@@ -5,8 +5,6 @@ import static pl.skidam.automodpack_core.Constants.*;
 import java.io.IOException;
 import java.net.ConnectException;
 import java.net.SocketTimeoutException;
-import java.nio.file.FileSystem;
-import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.LinkOption;
 import java.nio.file.Path;
@@ -16,7 +14,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -32,15 +29,12 @@ import pl.skidam.automodpack_core.modpack.generation.GenerationPatchNoteHistory;
 import pl.skidam.automodpack_core.modpack.generation.GenerationTarget;
 import pl.skidam.automodpack_core.modpack.generation.OwnershipLedger;
 import pl.skidam.automodpack_core.modpack.group.ClientPlatform;
-import pl.skidam.automodpack_core.modpack.group.ClientSelectionStore;
 import pl.skidam.automodpack_core.modpack.group.ResolvedSelection;
 import pl.skidam.automodpack_core.modpack.group.SelectedModpackTarget;
 import pl.skidam.automodpack_core.modpack.group.SelectionIntent;
 import pl.skidam.automodpack_core.protocol.DownloadClient;
 import pl.skidam.automodpack_core.update.ClientGenerationStore;
-import pl.skidam.automodpack_core.update.ClientOverlaySnapshot;
 import pl.skidam.automodpack_core.update.ClientStorage;
-import pl.skidam.automodpack_core.update.GeneratedCopyState;
 import pl.skidam.automodpack_core.update.RecoveryArchive;
 import pl.skidam.automodpack_core.update.UpdateDeferredException;
 import pl.skidam.automodpack_core.update.UpdatePlan;
@@ -50,13 +44,10 @@ import pl.skidam.automodpack_core.update.UpdateTransaction;
 import pl.skidam.automodpack_core.update.UpdateTransactionExecutor;
 import pl.skidam.automodpack_core.utils.DownloadSource;
 import pl.skidam.automodpack_core.utils.FetchManager;
-import pl.skidam.automodpack_core.utils.FileInspection;
-import pl.skidam.automodpack_core.utils.HashUtils;
 import pl.skidam.automodpack_core.utils.SmartFileUtils;
 import pl.skidam.automodpack_core.utils.UpdateLoopDetector;
 import pl.skidam.automodpack_core.utils.cache.FileMetadataCache;
 import pl.skidam.automodpack_core.utils.cache.ModFileCache;
-import pl.skidam.automodpack_core.utils.launchers.LauncherVersionSwapper;
 import pl.skidam.automodpack_loader_core.DetachedUpdateHelper;
 import pl.skidam.automodpack_loader_core.ReLauncher;
 import pl.skidam.automodpack_loader_core.UpdateTransactionSupport;
@@ -80,8 +71,9 @@ public class ModpackUpdater implements AutoCloseable {
 	private final AtomicReference<ConfirmationState> confirmationState = new AtomicReference<>(ConfirmationState.INACTIVE);
 	private final UpdateLoopDetector updateLoopDetector;
 	private final ClientStorage storage;
+	private final ClientUpdatePlanBuilder planBuilder;
 	private volatile FetchManager sourceFetchManager;
-	private PreparedPlan cachedSwitchPlan;
+	private ClientUpdatePlanBuilder.PreparedPlan cachedSwitchPlan;
 	private static final Comparator<RecoveryFile> RECOVERY_FILE_ORDER = Comparator.comparing(RecoveryFile::logicalPath).thenComparing(RecoveryFile::sha1)
 			.thenComparingLong(RecoveryFile::size);
 
@@ -135,9 +127,10 @@ public class ModpackUpdater implements AutoCloseable {
 		ClientStorageJsons.ClientGenerationStateFields active = storage.readActiveState();
 		if (active != null && selectedTarget.manifest().modpackId().equals(active.modpackId)) throw new IllegalArgumentException("Cached switch target is already active");
 		try (var cache = FileMetadataCache.open(storage.fileMetadataDirectory()); var modCache = ModFileCache.open(storage.modMetadataDirectory())) {
-			populateStoreFromCachedLocations(selectedTarget.flatTarget(), cache);
-			PreparedPlan prepared = buildPlan(cache, modCache, selectedTarget.flatTarget(), true);
-			ensurePlanObjects(prepared.plan(), selectedTarget.flatTarget());
+			planBuilder.populateStoreFromCachedLocations(selectedTarget.flatTarget(), cache);
+			ClientUpdatePlanBuilder.PreparedPlan prepared = planBuilder.buildPlan(
+					new ClientUpdatePlanBuilder.Input(selectedTarget, selectedTarget.flatTarget(), connectionInfo, clientConfig, true), cache, modCache);
+			planBuilder.ensurePlanObjects(prepared.plan(), selectedTarget.flatTarget());
 			cachedSwitchPlan = prepared;
 			List<GenerationPatchNoteHistory.Entry> missedPatchNotes = GenerationPatchNoteHistory.after(selectedTarget.patchNotesHistory(), "");
 			return UpdatePreview.create(prepared.plan(), prepared.originalFiles(), selectedTarget.flatTarget(), selectedTarget.selection(), false, null,
@@ -147,7 +140,7 @@ public class ModpackUpdater implements AutoCloseable {
 
 	/** Applies the last cached switch plan through the normal atomic transaction executor. */
 	public void applyCachedSwitch() throws Exception {
-		PreparedPlan prepared = cachedSwitchPlan;
+		ClientUpdatePlanBuilder.PreparedPlan prepared = cachedSwitchPlan;
 		if (prepared == null || selectedTarget == null) throw new IllegalStateException("Cached switch was not prepared");
 		try {
 			recordChangelogs(prepared, selectedTarget);
@@ -252,6 +245,7 @@ public class ModpackUpdater implements AutoCloseable {
 		this.serverModpackContent = selectedTarget == null ? null : selectedTarget.flatTarget();
 		this.connectionInfo = connectionInfo;
 		this.storage = Objects.requireNonNull(storage, "storage");
+		this.planBuilder = new ClientUpdatePlanBuilder(this.storage, MODPACK_LOADER, LOADER);
 		this.updateLoopDetector = new UpdateLoopDetector(storage.restartLoopStateFile());
 		this.downloadClient = downloadClient;
 	}
@@ -318,7 +312,7 @@ public class ModpackUpdater implements AutoCloseable {
 		try (var cache = FileMetadataCache.open(storage.fileMetadataDirectory())) {
 			Set<Jsons.ModpackContentFields.ModpackContentItem> allTargetItems = new LinkedHashSet<>(targetItems);
 			ModpackUtils.populateStoreFromCWD(allTargetItems, cache, storage);
-			populateStoreFromActive(preloadTarget, cache);
+			planBuilder.populateStoreFromActive(preloadTarget, cache);
 			Set<Jsons.ModpackContentFields.ModpackContentItem> targetSet = uniqueObjects(allTargetItems);
 			Set<Jsons.ModpackContentFields.ModpackContentItem> uncached = ModpackUtils.identifyUncachedFiles(targetSet, cache, storage);
 			if (uncached.isEmpty()) {
@@ -363,7 +357,8 @@ public class ModpackUpdater implements AutoCloseable {
 		if (selectedTarget == null || serverModpackContent == null) throw new IllegalStateException("Selected modpack target is unavailable");
 
 		try (var cache = FileMetadataCache.open(storage.fileMetadataDirectory()); var modCache = ModFileCache.open(storage.modMetadataDirectory())) {
-			PreparedPlan prepared = buildPlan(cache, modCache, selectedTarget.flatTarget(), false);
+			ClientUpdatePlanBuilder.PreparedPlan prepared = planBuilder.buildPlan(
+					new ClientUpdatePlanBuilder.Input(selectedTarget, selectedTarget.flatTarget(), connectionInfo, clientConfig, false), cache, modCache);
 			Jsons.ModpackContentFields installed = storedTarget();
 			return requiresReconciliation(prepared, installed);
 		}
@@ -371,14 +366,16 @@ public class ModpackUpdater implements AutoCloseable {
 
 	// Build the removal plan without changing the installed files.
 	public UpdatePreview previewRemoval() throws Exception {
-		RemovalPreparation preparation = prepareRemoval();
+		ClientUpdatePlanBuilder.RemovalPreparation preparation = planBuilder.prepareRemoval();
+		clientConfig = preparation.currentConfig();
 		return UpdatePreview.create(preparation.plan(), preparation.files(), preparation.installed(), removalSelection(preparation), true, preparation.baseline(),
 				preparation.completeFields().generation == null ? "" : preparation.completeFields().generation.patchNotes);
 	}
 
 	// Remove the installed modpack and restore baseline files before metadata cleanup.
 	public UpdateTransactionExecutor.Execution removeModpack() throws Exception {
-		RemovalPreparation preparation = prepareRemoval();
+		ClientUpdatePlanBuilder.RemovalPreparation preparation = planBuilder.prepareRemoval();
+		clientConfig = preparation.currentConfig();
 		UpdateTransaction transaction = UpdateTransaction.createRemoval(preparation.plan(), ClientPlatform.current(), preparation.expectedPriorIntent(), storage.overlayDigest(preparation.installed().modpackId));
 		UpdateTransactionExecutor.Execution execution = UpdateTransactionSupport.executor().commit(transaction);
 		if (execution.success()) {
@@ -392,48 +389,13 @@ public class ModpackUpdater implements AutoCloseable {
 		return execution;
 	}
 
-	private static ResolvedSelection removalSelection(RemovalPreparation preparation) {
+	private static ResolvedSelection removalSelection(ClientUpdatePlanBuilder.RemovalPreparation preparation) {
 		SelectionIntent intent = preparation.expectedPriorIntent();
 		if (intent == null) return null;
 		Set<String> selected = preparation.installed().selectedGroups == null ? Set.of() : preparation.installed().selectedGroups;
 		Set<String> stale = new TreeSet<>(intent.requestedGroups());
 		stale.removeAll(selected);
 		return new ResolvedSelection(intent, new TreeSet<>(selected), new TreeSet<>(stale));
-	}
-
-	private RemovalPreparation prepareRemoval() throws Exception {
-		Jsons.ModpackContentFields installed = storedTarget();
-		ClientStorageJsons.ClientGenerationStateFields activeState = storage.readActiveState();
-		if (activeState == null || !installed.modpackId.equals(activeState.modpackId)) throw new IOException("Active modpack generation state is missing");
-		Jsons.CompleteModpackContentFields completeFields = new ClientGenerationStore(storage).read(activeState.generationId)
-				.orElseThrow(() -> new IOException("Active client generation record is missing")).toFields();
-		ClientStorageJsons.ClientBaselineFields baseline = ConfigTools.read(storage.baselineFile(installed.modpackId), ClientStorageJsons.ClientBaselineFields.class)
-				.orElseGet(() -> {
-					ClientStorageJsons.ClientBaselineFields empty = new ClientStorageJsons.ClientBaselineFields();
-					empty.modpackId = installed.modpackId;
-					return empty;
-				});
-		Jsons.ClientConfigFieldsV3 currentConfig = ConfigTools.read(storage.clientConfigFile(), Jsons.ClientConfigFieldsV3.class)
-				.orElseGet(Jsons.ClientConfigFieldsV3::new);
-		Jsons.ClientConfigFieldsV3 plannedConfig = new Jsons.ClientConfigFieldsV3(currentConfig);
-		if (installed.modpackId.equals(plannedConfig.selectedModpackId)) plannedConfig.selectedModpackId = "";
-		clientConfig = currentConfig;
-		SelectionIntent expectedPriorIntent = new ClientSelectionStore(storage.selectionFile()).get(installed.modpackId).orElse(null);
-		GeneratedCopyState generatedCopies = expectedPriorIntent == null ? null : readGeneratedCopyState(installed, expectedPriorIntent);
-
-		try (var cache = FileMetadataCache.open(storage.fileMetadataDirectory())) {
-			Map<UpdatePlan.FileKey, UpdatePlan.FileState> files = inspectFiles(installed, installed, null,
-					generatedCopies == null ? List.of() : generatedCopies.nestedCopies(), cache,
-					Map.of(installed.modpackId, storage.overlaySnapshot(installed.modpackId, cache)));
-			Set<String> availableBaselineObjects = new HashSet<>();
-			if (baseline.entries != null) for (var entry : baseline.entries) {
-				if (entry == null || entry.absent || entry.objectHash == null || entry.size < 0) continue;
-				String hash = entry.objectHash.toLowerCase(Locale.ROOT);
-				if (SmartFileUtils.isValidFile(storage.objectsDirectory().resolve(hash), entry.size, hash)) availableBaselineObjects.add(hash);
-			}
-			UpdatePlan plan = UpdatePlanner.planRemoval(new UpdatePlanner.RemovalInput(installed, baseline, files, availableBaselineObjects, generatedCopies, plannedConfig));
-			return new RemovalPreparation(plan, completeFields, installed, baseline, expectedPriorIntent, plannedConfig, files);
-		}
 	}
 
 	// Load the already-installed modpack without contacting the server or
@@ -596,7 +558,7 @@ public class ModpackUpdater implements AutoCloseable {
 
 	private void startUpdateAfterPreview(Set<Jsons.ModpackContentFields.ModpackContentItem> filesToUpdate) {
 		long start = System.currentTimeMillis();
-		PreparedPlan finalPlan;
+		ClientUpdatePlanBuilder.PreparedPlan finalPlan;
 		try (var cache = FileMetadataCache.open(storage.fileMetadataDirectory()); var modCache = ModFileCache.open(storage.modMetadataDirectory())) {
 			requireLiveConnection();
 			// Don't download files which already exist
@@ -619,7 +581,7 @@ public class ModpackUpdater implements AutoCloseable {
 				throw e;
 			}
 
-			finalPlan = buildPlan(cache, modCache, selectedTarget.flatTarget());
+			finalPlan = planBuilder.buildPlan(new ClientUpdatePlanBuilder.Input(selectedTarget, selectedTarget.flatTarget(), connectionInfo, clientConfig, true), cache, modCache);
 		} catch (SocketTimeoutException | ConnectException e) {
 			String host = connectionInfo == null || connectionInfo.endpoint == null ? "modpack host" : "Modpack host of " + connectionInfo.endpoint.getHostString();
 			LOGGER.error("{} is not responding", host, e);
@@ -641,7 +603,7 @@ public class ModpackUpdater implements AutoCloseable {
 		applyApprovedPlan(finalPlan, start);
 	}
 
-	private ApplyStatus applyApprovedPlan(PreparedPlan plan, long start) {
+	private ApplyStatus applyApprovedPlan(ClientUpdatePlanBuilder.PreparedPlan plan, long start) {
 		try {
 			recordChangelogs(plan, selectedTarget);
 			ApplyResult applyResult = applyPreparedPlan(plan, selectedTarget);
@@ -757,7 +719,7 @@ public class ModpackUpdater implements AutoCloseable {
 	}
 
 	// this is run every time we modpack is updated
-	private ApplyResult applyPreparedPlan(PreparedPlan prepared, SelectedModpackTarget target) throws Exception {
+	private ApplyResult applyPreparedPlan(ClientUpdatePlanBuilder.PreparedPlan prepared, SelectedModpackTarget target) throws Exception {
 		executePlan(prepared, target);
 		UpdatePlan plan = prepared.plan();
 
@@ -769,7 +731,7 @@ public class ModpackUpdater implements AutoCloseable {
 		return result;
 	}
 
-	private void recordChangelogs(PreparedPlan prepared, SelectedModpackTarget target) {
+	private void recordChangelogs(ClientUpdatePlanBuilder.PreparedPlan prepared, SelectedModpackTarget target) {
 		UpdatePreview applied = UpdatePreview.create(prepared.plan(), prepared.originalFiles(), target.flatTarget(), target.selection(), false, null,
 				target.generationRecord().metadata().patchNotes(), target.patchNotesHistory());
 		Map<UpdatePlan.FileKey, List<String>> mainPageUrls = resolveMainPageUrls(prepared);
@@ -787,7 +749,7 @@ public class ModpackUpdater implements AutoCloseable {
 		LOGGER.info("Prepared update changes: {} changed, {} removed", changelogs.changedFiles().size(), changelogs.removedFiles().size());
 	}
 
-	private Map<UpdatePlan.FileKey, List<String>> resolveMainPageUrls(PreparedPlan prepared) {
+	private Map<UpdatePlan.FileKey, List<String>> resolveMainPageUrls(ClientUpdatePlanBuilder.PreparedPlan prepared) {
 		FetchManager manager = sourceFetchManager;
 		if (manager == null) return Map.of();
 		manager.fetch();
@@ -810,87 +772,11 @@ public class ModpackUpdater implements AutoCloseable {
 		return Map.copyOf(resolved);
 	}
 
-	private PreparedPlan buildPlan(FileMetadataCache cache, ModFileCache modCache, Jsons.ModpackContentFields target) throws Exception {
-		return buildPlan(cache, modCache, target, true);
-	}
-
-	private PreparedPlan buildPlan(FileMetadataCache cache, ModFileCache modCache, Jsons.ModpackContentFields target, boolean prepareObjects) throws Exception {
-		captureActiveEditableOverlays(cache);
-		Jsons.ModpackContentFields installed = storedTarget();
-		Map<String, ClientOverlaySnapshot> overlaySnapshots = new HashMap<>();
-		ClientOverlaySnapshot targetOverlay = storage.overlaySnapshot(target.modpackId, cache);
-		overlaySnapshots.put(target.modpackId, targetOverlay);
-		UpdatePlanner.SelectionContext selection = selectionContext(cache, overlaySnapshots);
-		GeneratedCopyState previousGeneratedState = installed == null ? null : readGeneratedCopyState(installed, selection);
-		Map<UpdatePlan.FileKey, UpdatePlan.FileState> files = inspectFiles(target, installed, selection,
-				previousGeneratedState == null ? List.of() : previousGeneratedState.nestedCopies(), cache, overlaySnapshots);
-		if (prepareObjects) populateStoreFromActive(target, cache);
-		Set<String> forceCopyServices = getForceCopyMods(target).stream().map(UpdatePlanner::normalize).collect(Collectors.toSet());
-		List<UpdatePlan.ModInfo> targetMods = inspectTargetMods(target, cache, modCache);
-		List<UpdatePlan.ModInfo> standardMods = inspectStandardMods(cache, modCache);
-		List<UpdatePlan.NestedCopy> nestedCopies = prepareObjects
-				? inspectNestedCopies(target, cache)
-				: readGeneratedCopyState(target, selectedTarget.selection().intent()).nestedCopies();
-		Jsons.ClientConfigFieldsV3 plannedConfig = connectionInfo == null || !connectionInfo.isComplete()
-				? ModpackUtils.planCachedModpackSelection(target.modpackId)
-				: ModpackUtils.planModpackSelection(target.modpackId, connectionInfo);
-		Map<String, UpdatePlan.FileState> editableOverlays = files.entrySet().stream().filter(entry -> entry.getKey().root() == UpdatePlan.Root.OVERLAY)
-				.collect(Collectors.toMap(entry -> entry.getKey().relativePath(), Map.Entry::getValue));
-
-		UpdatePlan plan = UpdatePlanner.plan(new UpdatePlanner.Input(installed, target, files, editableOverlays, forceCopyServices, targetMods, standardMods,
-				previousGeneratedState == null ? List.of() : previousGeneratedState.nestedCopies(), nestedCopies, selection, plannedConfig));
-		if (!LauncherVersionSwapper.requiresLoaderVersionSwap(target.loader, target.loaderVersion)) return new PreparedPlan(plan, files, targetOverlay.digest());
-		Set<UpdatePlan.RestartReason> restartReasons = EnumSet.noneOf(UpdatePlan.RestartReason.class);
-		restartReasons.addAll(plan.restartReasons());
-		restartReasons.add(UpdatePlan.RestartReason.CHANGED_LOADER_VERSION);
-		UpdatePlan withLoaderRestart = new UpdatePlan(plan.modpackId(), plan.generationTarget(), plan.operations(), plan.projectedFinalState(), plan.plannedClientConfig(),
-				restartReasons, plan.preservations(), plan.baselineCaptures(), plan.conflicts(), plan.generatedCopies());
-		return new PreparedPlan(withLoaderRestart, files, targetOverlay.digest());
-	}
-
-	private void captureActiveEditableOverlays(FileMetadataCache cache) throws IOException {
-		SelectedModpackTarget activeTarget = storedSelectedTarget();
-		if (activeTarget == null || activeTarget.flatTarget().list == null) return;
-		storage.ensureRoots();
-		Set<String> deletedPaths = new TreeSet<>(storage.readOverlayState(activeTarget.manifest().modpackId()).deletedPaths);
-		for (var item : activeTarget.flatTarget().list) {
-			if (!item.editable) continue;
-			Path live = livePath(item);
-			Path overlay = storage.overlayFile(activeTarget.manifest().modpackId(), item.file);
-			if (!Files.isRegularFile(live, LinkOption.NOFOLLOW_LINKS)) {
-				Files.deleteIfExists(overlay);
-				deletedPaths.add(UpdatePlanner.normalize(item.file));
-				continue;
-			}
-			String hash = cache.getHashOrNull(live);
-			if (hash == null) {
-				hash = HashUtils.getHash(live);
-				if (hash != null) cache.overwriteCache(live, hash);
-			}
-			if (hash == null) throw new IOException("Failed to hash editable live file: " + live);
-			long size = Files.size(live);
-			if (item.sha1.equalsIgnoreCase(hash) && Long.parseLong(item.size) == size) {
-				Files.deleteIfExists(overlay);
-				deletedPaths.remove(UpdatePlanner.normalize(item.file));
-				continue;
-			}
-			Path object = storage.objectsDirectory().resolve(hash);
-			if (!SmartFileUtils.isValidFile(object, size, hash)) SmartFileUtils.copyVerifiedAtomic(live, object, size, hash);
-			SmartFileUtils.copyVerifiedAtomic(object, overlay, size, hash);
-			deletedPaths.remove(UpdatePlanner.normalize(item.file));
-		}
-		storage.writeOverlayState(activeTarget.manifest().modpackId(), deletedPaths);
-	}
-
-	private Path livePath(Jsons.ModpackContentFields.ModpackContentItem item) {
-		String relative = UpdatePlanner.normalize(item.file);
-		return storage.gameDirectory().resolve(relative);
-	}
-
 	private PreviewRequestResult requestUpdatePreview(Set<Jsons.ModpackContentFields.ModpackContentItem> filesToUpdate) throws Exception {
 		if (selectedTarget == null) throw new IllegalStateException("Selected modpack target is unavailable");
 		try (var cache = FileMetadataCache.open(storage.fileMetadataDirectory()); var modCache = ModFileCache.open(storage.modMetadataDirectory())) {
-			PreparedPlan prepared = buildPlan(cache, modCache, selectedTarget.flatTarget(), false);
+			ClientUpdatePlanBuilder.PreparedPlan prepared = planBuilder.buildPlan(
+					new ClientUpdatePlanBuilder.Input(selectedTarget, selectedTarget.flatTarget(), connectionInfo, clientConfig, false), cache, modCache);
 			if (!requiresPlayerReview(prepared, firstConnection)) {
 				return switch (applyApprovedPlan(prepared, System.currentTimeMillis())) {
 					case APPLIED -> PreviewRequestResult.APPLIED;
@@ -913,23 +799,23 @@ public class ModpackUpdater implements AutoCloseable {
 	}
 
 	/** A review is required for the first install or any plan that changes player-visible state. */
-	private boolean requiresPlayerReview(PreparedPlan prepared, boolean firstInstall) {
+	private boolean requiresPlayerReview(ClientUpdatePlanBuilder.PreparedPlan prepared, boolean firstInstall) {
 		if (firstInstall) return true;
 		return hasPlanImpact(prepared);
 	}
 
 	/** Login reconciliation must also advance a newly advertised generation, even when its files are unchanged. */
-	private boolean requiresReconciliation(PreparedPlan prepared, Jsons.ModpackContentFields installed) {
+	private boolean requiresReconciliation(ClientUpdatePlanBuilder.PreparedPlan prepared, Jsons.ModpackContentFields installed) {
 		return installed == null || !GenerationTarget.fromFlat(installed).equals(prepared.plan().generationTarget()) || hasPlanImpact(prepared);
 	}
 
-	private boolean hasPlanImpact(PreparedPlan prepared) {
+	private boolean hasPlanImpact(ClientUpdatePlanBuilder.PreparedPlan prepared) {
 		UpdatePlan plan = prepared.plan();
 		return !plan.operations().isEmpty() || !plan.conflicts().isEmpty() || !plan.preservations().isEmpty() || !plan.baselineCaptures().isEmpty()
 				|| !plan.restartReasons().isEmpty() || !ConfigTools.GSON.toJson(plan.plannedClientConfig()).equals(ConfigTools.GSON.toJson(clientConfig));
 	}
 
-	private boolean requestPreparedPlanPreview(PreparedPlan prepared, Runnable continueAction, Runnable cancelAction, boolean returnToSelection) throws IOException {
+	private boolean requestPreparedPlanPreview(ClientUpdatePlanBuilder.PreparedPlan prepared, Runnable continueAction, Runnable cancelAction, boolean returnToSelection) throws IOException {
 		List<GenerationPatchNoteHistory.Entry> missedPatchNotes = GenerationPatchNoteHistory.after(selectedTarget.patchNotesHistory(), installedGenerationId());
 		UpdatePreview preview = UpdatePreview.create(prepared.plan(), prepared.originalFiles(), selectedTarget.flatTarget(), selectedTarget.selection(), false, null, getPatchNotes(), missedPatchNotes);
 		return new ScreenManager().preview(preview, getModpackName(),
@@ -941,182 +827,9 @@ public class ModpackUpdater implements AutoCloseable {
 		return state != null && selectedTarget != null && selectedTarget.manifest().modpackId().equals(state.modpackId) ? state.generationId : "";
 	}
 
-	private UpdatePlanner.SelectionContext selectionContext(FileMetadataCache cache, Map<String, ClientOverlaySnapshot> overlaySnapshots) throws IOException {
-		String previousId = clientConfig.selectedModpackId;
-		if (previousId == null || previousId.isBlank() || !ModpackId.isValid(previousId)) return null;
-		Jsons.ModpackContentFields previousManifest = storedTarget();
-		ClientOverlaySnapshot snapshot = overlaySnapshots.get(previousId);
-		if (snapshot == null) {
-			snapshot = storage.overlaySnapshot(previousId, cache);
-			overlaySnapshots.put(previousId, snapshot);
-		}
-		return new UpdatePlanner.SelectionContext(previousId, previousManifest, snapshot.files());
-	}
-
-	private void populateStoreFromActive(Jsons.ModpackContentFields target, FileMetadataCache cache) throws IOException {
-		populateStoreFromSources(target, cache, item -> List.of(SmartFileUtils.getPath(storage.activeDirectory(), item.file)));
-	}
-
-	private void populateStoreFromCachedLocations(Jsons.ModpackContentFields target, FileMetadataCache cache) throws IOException {
-		populateStoreFromSources(target, cache,
-				item -> List.of(SmartFileUtils.getPath(storage.activeDirectory(), item.file), livePath(item)));
-	}
-
-	private void populateStoreFromSources(Jsons.ModpackContentFields target, FileMetadataCache cache,
-			Function<Jsons.ModpackContentFields.ModpackContentItem, List<Path>> sourceResolver) throws IOException {
-		if (target.list == null) return;
-		for (var item : target.list) {
-			Path object = storage.objectsDirectory().resolve(item.sha1);
-			long size = Long.parseLong(item.size);
-			if (SmartFileUtils.isValidFile(object, size, item.sha1)) continue;
-			for (Path source : sourceResolver.apply(item))
-				if (populateStoreObject(source, object, size, item.sha1, cache)) break;
-		}
-	}
-
-	private static boolean populateStoreObject(Path source, Path object, long size, String sha1, FileMetadataCache cache) throws IOException {
-		if (!SmartFileUtils.isValidFile(source, size, sha1)) return false;
-		SmartFileUtils.copyVerifiedAtomic(source, object, size, sha1);
-		cache.overwriteCache(object, sha1);
-		return true;
-	}
-
-	private Map<UpdatePlan.FileKey, UpdatePlan.FileState> inspectFiles(Jsons.ModpackContentFields target, Jsons.ModpackContentFields installed,
-			UpdatePlanner.SelectionContext selection, List<UpdatePlan.NestedCopy> previousGeneratedCopies, FileMetadataCache cache,
-			Map<String, ClientOverlaySnapshot> overlaySnapshots) throws IOException {
-		Map<UpdatePlan.FileKey, UpdatePlan.FileState> files = new HashMap<>();
-		if (Files.isDirectory(storage.activeDirectory())) {
-			try (Stream<Path> stream = Files.walk(storage.activeDirectory())) {
-				for (Path path : stream.filter(path -> Files.isRegularFile(path, LinkOption.NOFOLLOW_LINKS)).toList())
-					putFileState(files, UpdatePlan.Root.PROJECTION, storage.activeDirectory(), path, cache);
-			}
-		}
-		ClientOverlaySnapshot targetOverlay = overlaySnapshots.get(target.modpackId);
-		if (targetOverlay == null) {
-			targetOverlay = storage.overlaySnapshot(target.modpackId, cache);
-			overlaySnapshots.put(target.modpackId, targetOverlay);
-		}
-		for (var entry : targetOverlay.files().entrySet())
-			files.put(new UpdatePlan.FileKey(UpdatePlan.Root.OVERLAY, entry.getKey()), entry.getValue());
-		Set<String> gamePaths = new HashSet<>();
-		if (target.list != null) target.list.forEach(item -> gamePaths.add(item.file));
-		if (installed != null && installed.list != null) installed.list.forEach(item -> gamePaths.add(item.file));
-		if (selection != null && selection.previousManifest() != null && selection.previousManifest().list != null) selection.previousManifest().list.forEach(item -> gamePaths.add(item.file));
-		previousGeneratedCopies.forEach(copy -> gamePaths.add(copy.relativePath()));
-		if (installed != null && installed.ownershipLedger != null && installed.ownershipLedger.entries != null)
-			installed.ownershipLedger.entries.forEach(entry -> {
-				if (entry != null) gamePaths.add(entry.logicalPath);
-			});
-		for (String gamePath : gamePaths) {
-			Path path = SmartFileUtils.getPath(storage.gameDirectory(), gamePath);
-			if (Files.isRegularFile(path, LinkOption.NOFOLLOW_LINKS)) putFileState(files, UpdatePlan.Root.GAME_DIR, storage.gameDirectory(), path, cache);
-		}
-		if (Files.isDirectory(storage.modsDirectory(), LinkOption.NOFOLLOW_LINKS)) {
-			try (Stream<Path> stream = Files.list(storage.modsDirectory())) {
-				for (Path path : stream.filter(candidate -> Files.isRegularFile(candidate, LinkOption.NOFOLLOW_LINKS)).toList())
-					putFileState(files, UpdatePlan.Root.GAME_DIR, storage.gameDirectory(), path, cache);
-			}
-		}
-		OwnershipLedger ledger = OwnershipLedger.fromFields(target.ownershipLedger);
-		for (String logicalPath : ledger.entries().keySet()) {
-			Optional<UpdatePlan.FileKey> cleanupKey = UpdatePlanner.managedCleanupKey(logicalPath);
-			if (cleanupKey.isEmpty()) continue;
-			UpdatePlan.FileKey key = cleanupKey.get();
-			if (key.root() != UpdatePlan.Root.GAME_DIR) continue;
-			Path path = storage.gameDirectory().resolve(key.relativePath()).normalize();
-			if (path.startsWith(storage.gameDirectory()) && Files.isRegularFile(path, LinkOption.NOFOLLOW_LINKS)) putFileState(files, key.root(), storage.gameDirectory(), path, cache);
-		}
-		return files;
-	}
-
-	private GeneratedCopyState readGeneratedCopyState(Jsons.ModpackContentFields manifest, UpdatePlanner.SelectionContext selection) throws IOException {
-		SelectionIntent intent = selection == null ? null : new ClientSelectionStore(storage.selectionFile()).get(manifest.modpackId).orElse(null);
-		return intent == null ? null : readGeneratedCopyState(manifest, intent);
-	}
-
-	private GeneratedCopyState readGeneratedCopyState(Jsons.ModpackContentFields manifest, SelectionIntent intent) throws IOException {
-		String digest = UpdateTransaction.digest(intent);
-		if (digest.isEmpty()) throw new IOException("Cannot read generated-copy state without a selected group intent");
-		return GeneratedCopyState.read(storage, manifest.modpackId, manifest.targetGenerationId, digest);
-	}
-
-	private void putFileState(Map<UpdatePlan.FileKey, UpdatePlan.FileState> files, UpdatePlan.Root root, Path rootPath, Path path,
-			FileMetadataCache cache) throws IOException {
-		if (!Files.isRegularFile(path, LinkOption.NOFOLLOW_LINKS)) return;
-		String relative = UpdatePlanner.normalize(rootPath.toAbsolutePath().normalize().relativize(path.toAbsolutePath().normalize()).toString());
-		String hash = cache.getHashOrNull(path);
-		if (hash == null) {
-			hash = HashUtils.getHash(path);
-			if (hash == null) throw new IOException("Failed to hash live file: " + path);
-			cache.overwriteCache(path, hash);
-		}
-		files.put(new UpdatePlan.FileKey(root, relative), new UpdatePlan.FileState(hash, Files.size(path), true, FileInspection.isMod(path)));
-	}
-
-	private List<UpdatePlan.ModInfo> inspectTargetMods(Jsons.ModpackContentFields target, FileMetadataCache cache, ModFileCache modCache) {
-		List<UpdatePlan.ModInfo> mods = new ArrayList<>();
-		for (var item : target.list.stream().filter(value -> "mod".equals(value.type)).sorted(Comparator.comparing(value -> value.file)).toList()) {
-			long size = Long.parseLong(item.size);
-			Path source = storage.objectsDirectory().resolve(item.sha1);
-			if (!SmartFileUtils.isValidFile(source, size, item.sha1)) source = SmartFileUtils.getPath(storage.activeDirectory(), item.file);
-			if (!SmartFileUtils.isValidFile(source, size, item.sha1)) continue;
-			FileInspection.Mod mod = modCache.getModOrNull(source, cache);
-			if (mod != null) mods.add(new UpdatePlan.ModInfo(UpdatePlanner.normalize(item.file), item.sha1, size, mod.IDs(), mod.deps()));
-		}
-		return mods;
-	}
-
-	private List<UpdatePlan.ModInfo> inspectStandardMods(FileMetadataCache cache, ModFileCache modCache) throws IOException {
-		if (!Files.isDirectory(storage.modsDirectory())) return List.of();
-		List<UpdatePlan.ModInfo> mods = new ArrayList<>();
-		try (Stream<Path> stream = Files.list(storage.modsDirectory())) {
-			for (Path path : stream.filter(Files::isRegularFile).sorted().toList()) {
-				FileInspection.Mod mod = modCache.getModOrNull(path, cache);
-				if (mod != null) {
-					String relativePath = UpdatePlanner.normalize(storage.gameDirectory().relativize(path.toAbsolutePath().normalize()).toString());
-					mods.add(new UpdatePlan.ModInfo(relativePath, mod.hash(), Files.size(path), mod.IDs(), mod.deps()));
-				}
-			}
-		}
-		return mods;
-	}
-
-	private List<UpdatePlan.NestedCopy> inspectNestedCopies(Jsons.ModpackContentFields target, FileMetadataCache cache) throws IOException {
-		Files.createDirectories(storage.incomingDirectory());
-		Path inspectionDirectory = Files.createTempDirectory(storage.incomingDirectory(), "inspection-");
-		try {
-			Path inspectionMods = inspectionDirectory.resolve("mods");
-			Files.createDirectories(inspectionMods);
-			for (var item : target.list.stream().filter(value -> "mod".equals(value.type)).toList()) {
-				Path source = storage.objectsDirectory().resolve(item.sha1);
-				if (!SmartFileUtils.isValidFile(source, Long.parseLong(item.size), item.sha1)) source = SmartFileUtils.getPath(storage.activeDirectory(), item.file);
-				if (!SmartFileUtils.isValidFile(source, Long.parseLong(item.size), item.sha1)) continue;
-				SmartFileUtils.copyVerifiedAtomic(source, inspectionMods.resolve(Path.of(UpdatePlanner.normalize(item.file)).getFileName()), Long.parseLong(item.size),
-						item.sha1);
-			}
-
-			List<UpdatePlan.NestedCopy> copies = new ArrayList<>();
-			for (FileInspection.Mod mod : MODPACK_LOADER.getModpackNestedConflicts(inspectionDirectory, cache)) {
-				if (mod.path() == null || mod.hash() == null || !Files.isRegularFile(mod.path())) continue;
-				long size = Files.size(mod.path());
-				Path storeFile = storage.objectsDirectory().resolve(mod.hash());
-				if (!SmartFileUtils.isValidFile(storeFile, size, mod.hash())) SmartFileUtils.copyVerifiedAtomic(mod.path(), storeFile, size, mod.hash());
-				Path targetPath = storage.modsDirectory().resolve(mod.path().getFileName()).normalize();
-				if (!targetPath.startsWith(storage.gameDirectory())) throw new IOException("Nested mod target escaped the game directory: " + targetPath);
-				String relativePath = UpdatePlanner.normalize(storage.gameDirectory().relativize(targetPath).toString());
-				copies.add(new UpdatePlan.NestedCopy(relativePath, mod.hash(), size, mod.IDs()));
-			}
-			return copies;
-		} finally {
-			try (Stream<Path> stream = Files.walk(inspectionDirectory)) {
-				for (Path path : stream.sorted(Comparator.reverseOrder()).toList()) Files.deleteIfExists(path);
-			}
-		}
-	}
-
-	private void executePlan(PreparedPlan prepared, SelectedModpackTarget target) throws IOException {
+	private void executePlan(ClientUpdatePlanBuilder.PreparedPlan prepared, SelectedModpackTarget target) throws IOException {
 		UpdatePlan plan = prepared.plan();
-		ensurePlanObjects(plan, target.flatTarget());
+		planBuilder.ensurePlanObjects(plan, target.flatTarget());
 		UpdateTransactionExecutor.Execution execution = UpdateTransactionSupport.executor().commit(plan, target, prepared.overlayDigest());
 		if (!execution.success()) {
 			DetachedUpdateHelper.launch(execution.transaction());
@@ -1143,45 +856,6 @@ public class ModpackUpdater implements AutoCloseable {
 			if (operation.root() == UpdatePlan.Root.OVERLAY && operation.operation() == UpdatePlan.OperationType.DELETE)
 				deletedPaths.remove(UpdatePlanner.normalize(operation.relativePath()));
 		storage.writeOverlayState(modpackId, deletedPaths);
-	}
-
-	private void populateStoreFromModpack(Collection<Jsons.ModpackContentFields.ModpackContentItem> items, FileMetadataCache cache) {
-		for (var item : items) {
-			Path storeFile = storage.objectsDirectory().resolve(item.sha1);
-			long size = Long.parseLong(item.size);
-			if (SmartFileUtils.isValidFile(storeFile, size, item.sha1)) continue;
-			Path source = SmartFileUtils.getPath(storage.activeDirectory(), item.file);
-			try {
-				populateStoreObject(source, storeFile, size, item.sha1, cache);
-			} catch (IOException e) {
-				LOGGER.warn("Failed to reuse installed modpack object {}", item.file, e);
-			}
-		}
-	}
-
-	private void ensurePlanObjects(UpdatePlan plan, Jsons.ModpackContentFields targetManifest) throws IOException {
-		Map<String, Jsons.ModpackContentFields.ModpackContentItem> itemsByHash = targetManifest.list.stream()
-				.collect(Collectors.toMap(item -> item.sha1.toLowerCase(Locale.ROOT), item -> item, (first, second) -> first));
-		for (UpdatePlan.Operation operation : plan.operations()) {
-			if (operation.operation() != UpdatePlan.OperationType.INSTALL_OBJECT) continue;
-			Path storeFile = storage.objectsDirectory().resolve(operation.expectedObjectHash());
-			if (SmartFileUtils.isValidFile(storeFile, operation.expectedSize(), operation.expectedObjectHash())) continue;
-			if (operation.root() == UpdatePlan.Root.OVERLAY) {
-				Path overlay = storage.overlayFile(targetManifest.modpackId, operation.relativePath());
-				if (SmartFileUtils.isValidFile(overlay, operation.expectedSize(), operation.expectedObjectHash())) {
-					SmartFileUtils.copyVerifiedAtomic(overlay, storeFile, operation.expectedSize(), operation.expectedObjectHash());
-					continue;
-				}
-				throw new IOException("Required editable overlay object is unavailable: " + operation.expectedObjectHash());
-			}
-			var item = itemsByHash.get(operation.expectedObjectHash().toLowerCase(Locale.ROOT));
-			if (item == null) throw new IOException("Planned CAS object is unavailable: " + operation.expectedObjectHash());
-			Path source = SmartFileUtils.getPath(storage.activeDirectory(), item.file);
-			if (!SmartFileUtils.isValidFile(source, operation.expectedSize(), operation.expectedObjectHash())) source = livePath(item);
-			if (!SmartFileUtils.isValidFile(source, operation.expectedSize(), operation.expectedObjectHash()))
-				throw new IOException("Required object is absent from CAS and verified live locations: " + operation.expectedObjectHash());
-			SmartFileUtils.copyVerifiedAtomic(source, storeFile, operation.expectedSize(), operation.expectedObjectHash());
-		}
 	}
 
 	private boolean beginConfirmation() {
@@ -1221,21 +895,6 @@ public class ModpackUpdater implements AutoCloseable {
 		}
 	}
 
-	private record PreparedPlan(UpdatePlan plan, Map<UpdatePlan.FileKey, UpdatePlan.FileState> originalFiles, String overlayDigest) {
-		private PreparedPlan {
-			originalFiles = Map.copyOf(originalFiles);
-			if (overlayDigest == null || !overlayDigest.matches("[0-9a-f]{40}")) throw new IllegalArgumentException("Prepared overlay digest is invalid");
-		}
-	}
-
-	private record RemovalPreparation(UpdatePlan plan, Jsons.CompleteModpackContentFields completeFields, Jsons.ModpackContentFields installed,
-			ClientStorageJsons.ClientBaselineFields baseline, SelectionIntent expectedPriorIntent, Jsons.ClientConfigFieldsV3 plannedConfig,
-			Map<UpdatePlan.FileKey, UpdatePlan.FileState> files) {
-		private RemovalPreparation {
-			files = Map.copyOf(files);
-		}
-	}
-
 	private record ApplyResult(EnumSet<RestartReason> restartReasons) {
 		private boolean requiresRestart() {
 			return !restartReasons.isEmpty();
@@ -1258,26 +917,4 @@ public class ModpackUpdater implements AutoCloseable {
 		PREVIEW_SHOWN, PREVIEW_NOT_SHOWN, APPLIED, DEFERRED, FAILED
 	}
 
-	// Returns the modpack mods that ship a service file this loader's running version cannot host
-	// in place (see ModpackLoaderService#forceCopyServices) - these must be copied into standard
-	// mods/ instead of staying in the active projection.
-	private Set<String> getForceCopyMods(Jsons.ModpackContentFields modpackContentFields) throws IOException {
-		Set<String> forceCopyServices = MODPACK_LOADER.forceCopyServices();
-		Set<String> forceCopyMods = new HashSet<>();
-		if (forceCopyServices.isEmpty()) return forceCopyMods;
-
-		for (Jsons.ModpackContentFields.ModpackContentItem item : modpackContentFields.list) {
-			if (!item.type.equals("mod")) continue;
-
-			long size = Long.parseLong(item.size);
-			Path modPath = storage.objectsDirectory().resolve(item.sha1);
-			if (!SmartFileUtils.isValidFile(modPath, size, item.sha1)) modPath = SmartFileUtils.getPath(storage.activeDirectory(), item.file);
-			if (!SmartFileUtils.isValidFile(modPath, size, item.sha1)) continue;
-			try (FileSystem fs = FileSystems.newFileSystem(modPath)) {
-				if (!FileInspection.getServices(fs, forceCopyServices).isEmpty()) forceCopyMods.add(item.file);
-			}
-		}
-
-		return forceCopyMods;
-	}
 }
